[MOV] website_sale_stock(_wishlist->{}): merge wishlist
