+++ .weblate.json
@@ -1732,11 +1732,6 @@
                 "new_base": "addons/website_sale_stock/i18n/website_sale_stock.pot"
             },
             {
-                "name": "website_sale_stock_wishlist",
-                "filemask": "addons/website_sale_stock_wishlist/i18n/*.po",
-                "new_base": "addons/website_sale_stock_wishlist/i18n/website_sale_stock_wishlist.pot"
-            },
-            {
                 "name": "website_slides",
                 "filemask": "addons/website_slides/i18n/*.po",
                 "new_base": "addons/website_slides/i18n/website_slides.pot"
