+++ addons/website_sale_stock/__manifest__.py
@@ -19,6 +19,7 @@ Then it can be made specific at the product level.
         'views/product_template_views.xml',
         'views/res_config_settings_views.xml',
         'views/website_sale_stock_templates.xml',
+        'views/wishlist_templates.xml',
         'data/template_email.xml',
         'data/ir_cron_data.xml',
     ],
@@ -35,6 +36,7 @@ Then it can be made specific at the product level.
             'website_sale_stock/static/src/js/product_card/*',
             'website_sale_stock/static/src/js/product_configurator_dialog/*',
             'website_sale_stock/static/src/xml/**/*',
+            'website_sale_stock/static/src/scss/**/*',
         ],
         'web.assets_tests': [
             'website_sale_stock/static/tests/tours/*',
