+++ addons/website_sale_stock/controllers/variant.py
@@ -10,4 +10,5 @@ class WebsiteSaleStockVariantController(WebsiteSaleVariantController):
     @route()
     def get_combination_info_website(self, *args, **kwargs):
         request.update_context(website_sale_stock_get_quantity=True)
+        request.update_context(website_sale_stock_wishlist_get_wish=True)
         return super().get_combination_info_website(*args, **kwargs)
