+++ addons/website_sale_stock/models/__init__.py
@@ -1,13 +1,15 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import website
-from . import product_combo
-from . import product_feed
-from . import product_product
-from . import product_ribbon
-from . import product_template
-from . import res_config_settings
-from . import sale_order
-from . import sale_order_line
-from . import stock_picking
+from . import (
+    product_combo,
+    product_feed,
+    product_product,
+    product_ribbon,
+    product_template,
+    product_wishlist,
+    res_config_settings,
+    sale_order,
+    sale_order_line,
+    stock_picking,
+    website,
+)
