+++ addons/website_sale_stock/models/product_template.py
@@ -104,6 +104,13 @@ class ProductTemplate(models.Model):
                 'cart_qty': 0,
             })
 
+        if not self.env.context.get('website_sale_stock_wishlist_get_wish'):
+            return res
+
+        if product_or_template.is_product_variant:
+            product_sudo = product_or_template.sudo()
+            res['is_in_wishlist'] = product_sudo._is_in_wishlist()
+
         return res
 
     @api.model
