--- addons/website_sale_stock_wishlist/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import models
-from . import controllers
