--- addons/website_sale_stock_wishlist/i18n/ru.po
@@ -1,66 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* website_sale_stock_wishlist
-#
-# Translators:
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 17.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-12-16 18:37+0000\n"
-"PO-Revision-Date: 2024-01-30 15:14+0400\n"
-"Last-Translator: \n"
-"Language-Team: Russian (https://app.transifex.com/odoo/teams/41243/ru/)\n"
-"Language: ru\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=4; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && "
-"n%10<=4 && (n%100<12 || n%100>14) ? 1 : n%10==0 || (n%10>=5 && n%10<=9) || "
-"(n%100>=11 && n%100<=14)? 2 : 3);\n"
-
-#. module: website_sale_stock_wishlist
-#: model_terms:ir.ui.view,arch_db:website_sale_stock_wishlist.product_wishlist
-msgid ""
-"<i class=\"fa fa-circle text-danger me-1\" aria-hidden=\"true\"/>\n"
-"                    Out of stock"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#. odoo-javascript
-#: code:addons/website_sale_stock_wishlist/static/src/xml/product_availability.xml:0
-msgid "Add to wishlist"
-msgstr "Добавить в избранное"
-
-#. module: website_sale_stock_wishlist
-#. odoo-javascript
-#: code:addons/website_sale_stock_wishlist/static/src/xml/product_availability.xml:0
-msgid "Added to your wishlist"
-msgstr "Добавлено в ваш список желаний"
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_template__display_name
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_template__id
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__id
-msgid "ID"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model,name:website_sale_stock_wishlist.model_product_template
-msgid "Product"
-msgstr "Товар"
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model,name:website_sale_stock_wishlist.model_product_wishlist
-msgid "Product Wishlist"
-msgstr "Избранные товары"
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__stock_notification
-msgid "Stock Notification"
-msgstr "Уведомление о запасах"
