--- addons/website_sale_stock_wishlist/i18n/zh_TW.po
@@ -1,69 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# * website_sale_stock_wishlist
-#
-# Translators:
-# Wil Odoo, 2025
-#
-# "Dylan Kiss (dyki)" <dyki@odoo.com>, 2025.
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server saas~18.3\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-12-16 18:37+0000\n"
-"PO-Revision-Date: 2025-09-16 08:12+0000\n"
-"Last-Translator: \"Dylan Kiss (dyki)\" <dyki@odoo.com>\n"
-"Language-Team: Chinese (Traditional Han script) <https://translate.odoo.com/"
-"projects/odoo-19/website_sale_stock_wishlist/zh_Hant/>\n"
-"Language: zh_TW\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-"X-Generator: Weblate 5.12.2\n"
-
-#. module: website_sale_stock_wishlist
-#: model_terms:ir.ui.view,arch_db:website_sale_stock_wishlist.product_wishlist
-msgid ""
-"<i class=\"fa fa-circle text-danger me-1\" aria-hidden=\"true\"/>\n"
-"                    Out of stock"
-msgstr ""
-
-#. module: website_sale_stock_wishlist
-#. odoo-javascript
-#: code:addons/website_sale_stock_wishlist/static/src/xml/product_availability.xml:0
-msgid "Add to wishlist"
-msgstr "加入願望清單"
-
-#. module: website_sale_stock_wishlist
-#. odoo-javascript
-#: code:addons/website_sale_stock_wishlist/static/src/xml/product_availability.xml:0
-msgid "Added to your wishlist"
-msgstr "已加至願望清單"
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_template__display_name
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__display_name
-msgid "Display Name"
-msgstr "顯示名稱"
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_template__id
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__id
-msgid "ID"
-msgstr "識別號"
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model,name:website_sale_stock_wishlist.model_product_template
-msgid "Product"
-msgstr "商品"
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model,name:website_sale_stock_wishlist.model_product_wishlist
-msgid "Product Wishlist"
-msgstr "產品願望清單"
-
-#. module: website_sale_stock_wishlist
-#: model:ir.model.fields,field_description:website_sale_stock_wishlist.field_product_wishlist__stock_notification
-msgid "Stock Notification"
-msgstr "到貨通知"
