--- addons/website_sale_stock_wishlist/models/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import product_template
-from . import product_wishlist
